+++ flutter_packages/quran_library/lib/quran.dart
@@ -1,4 +1,4 @@
-import 'dart:async' show Completer, StreamSubscription, Timer;
+import 'dart:async' show Completer, StreamSubscription, Timer, unawaited;
 import 'dart:convert' show jsonDecode, jsonEncode;
 import 'dart:developer' show log;
 import 'dart:math' as math show max, min;